Move asn database to new package
