+++ .gitignore
@@ -1,1 +1,2 @@
-flow-exporter
+./flow-exporter
+dist/
